# src/main/resources/openapi/api-specification.yaml
openapi: 3.0.0
info:
  title: Technical Challengae
  description: Backend Service
  version: 1.0.0
  x-audience: external-public
  x-api-id: 78e624f8-f73e-4bbe-a918-bda261fd13ec
  contact:
    name: zatec
    url: http://www.zatec.com
    email: info@zatec.com
components:
  responses:
...
    JokeStatusResponse:
      description: this contains random joke
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JokeResponse"
  schemas:
...
    JokeResponse:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: "#/components/schemas/Jokes"
    Jokes:
      type: object
      properties:
        categories:
          type: array
          items:
            type: string
        created_at:
          type: string
        icon_url:
          type: string
        id:
          type: string
        updated_at:
          type: string
        url:
          type: string
        value:
          type: string
...
servers:
  - url: 'http://localhost:8080'
    description: locally served url
...
paths:
...
  /chuck/random:
    get:
      summary: Get random joke by category
      description: get random joke
      operationId: getRandomJoke
      tags:
        - chuck
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
      responses:
        '200':
          $ref: "#/components/responses/JokeStatusResponse"